Moved aws models
