+++ server/src/main/java/com/omigost/server/iam/model/AwsUser.java
@@ -1,4 +1,4 @@
-package com.omigost.server.model;
+package com.omigost.server.iam.model;
 
 import lombok.Builder;
 import lombok.Data;
