--- server/src/main/java/com/omigost/server/model/Limit.java
@@ -1,4 +1,0 @@
-package com.omigost.server.model;
-
-public class Limit {
-}
