+++ server/src/main/java/com/omigost/server/rest/AWSTestController.java
@@ -1,7 +1,7 @@
 package com.omigost.server.rest;
 
 import com.omigost.server.iam.IAMService;
-import com.omigost.server.model.AwsUser;
+import com.omigost.server.iam.model.AwsUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
